fix: don't declare constructor/destructor at header file
ArrayType in type.h contains opqeue type Expression as unique_ptr, and
it cause error when type.h is included without expr.h.
It happen because unique_ptr require complete type for its destructor,
and compiler uses it in automatically generated destructor and maybe in
constructor defined by me.
So, I remove definition from header file and instead implement it in
source file to solve it.