fix: incorrect code generation
As struct or array is normally big compare to builtin types, so we
introduced fat object. If the value is fat object, we don't pushes
entire contents, and instead of this, we pushes addres of the value.

Previous code pushes address when struct or array object is generated,
but don't treat it when copy it. So fix the codes.

Also, the use of mulq is incorrect and cause assemble error, fix it.